Kcal Float instead of Integer

+++ forms.py
@@ -14,7 +14,7 @@ class AddRecipe(FlaskForm):
 
 class AddProduct(FlaskForm):
   title =  StringField("Name of the product", validators=[DataRequired()])
-  calories = IntegerField("Calories per 100 gr.", default=1)
+  calories = FloatField("Calories per 100 gr.", default=1)
   proteins = FloatField("Proteins per 100 gr.", default=1)
   check = SubmitField("Check")
   submit = SubmitField("Confirm!")
